refactored bids parser
